fix(tcp): make active_connections atomic to prevent deinit hang and use-after-free
tcp_desc_t::active_connections was a plain int modified concurrently by
the acceptor task (++) and per-client receiver tasks (--) on dual-core
ESP32 without any synchronization. This is a real parallel data race:

- Lost decrement keeps counter > 0 → tcp_server_deinit spins forever
- Lost increment allows free(desc) while a receiver task still runs
  → use-after-free

Fix: change field type to volatile uint32_t and protect all
read-modify-write sites with Atomic_Increment_u32 / Atomic_Decrement_u32
from freertos/atomic.h (implemented via ATOMIC_ENTER/EXIT_CRITICAL).

In tcp_client.c the reset after disconnect uses a plain store to 0
(not Atomic_Decrement_u32) because there is exactly one writer task
in client mode; volatile guarantees compiler-level visibility and
ESP32 shared SRAM provides hardware-level visibility without a barrier.

Also fix implicit narrowing uint32_t → int in bridge.c by adding an
explicit (int) cast on the return from tcp_server_active_connections().

Add unittests/mocks/freertos/atomic.h implementing the same API via
GCC __atomic_fetch_add/__atomic_fetch_sub for unit-test builds.

Files changed:
- main/bridge/tcp_desc.h: int → volatile uint32_t
- main/bridge/tcp_server.c: 4 sites → Atomic_Increment/Decrement_u32
- main/bridge/tcp_client.c: ++ → Atomic_Increment_u32; = 0 kept as plain store
- main/bridge/bridge.c: add (int) cast
- unittests/mocks/freertos/atomic.h: new mock file

+++ main/bridge/bridge.c
@@ -76,7 +76,7 @@ int tcp_server_active_connections(tcp_server_num_t server_num)
         return 0;
     }
 
-    return bridge_ctx[server_num].tcp_desc->active_connections;
+    return (int)bridge_ctx[server_num].tcp_desc->active_connections;
 }
 
 esp_err_t bridge_disable_port(unsigned index)
